refactor(services): move some services into sub-folders
