+++ src/features/discord/messages/services/command/cookie/discord-message-command-cookie.service.spec.ts
@@ -1,0 +1,321 @@
+import moment from "moment-timezone";
+import { createMock } from "ts-auto-mock";
+import { ColorEnum } from "../../../../../../enums/color.enum";
+import { IconEnum } from "../../../../../../enums/icon.enum";
+import { ServiceNameEnum } from "../../../../../../enums/service-name.enum";
+import { CoreEventService } from "../../../../../core/services/core-event.service";
+import { ILoggerLog } from "../../../../../logger/interfaces/logger-log";
+import { LoggerService } from "../../../../../logger/services/logger.service";
+import { DiscordSoniaService } from "../../../../users/services/discord-sonia.service";
+import { DiscordMessageCommandCookieDescriptionEnum } from "../../../enums/command/cookie/discord-message-command-cookie-description.enum";
+import { DiscordMessageCommandCookieTitleEnum } from "../../../enums/command/cookie/discord-message-command-cookie-title.enum";
+import { AnyDiscordMessage } from "../../../types/any-discord-message";
+import { DiscordMessageConfigService } from "../../config/discord-message-config.service";
+import { DiscordMessageCommandCookieService } from "./discord-message-command-cookie.service";
+import {
+  MessageEmbedAuthor,
+  MessageEmbedFooter,
+  MessageEmbedThumbnail,
+} from "discord.js";
+import * as GetRandomValueFromEnumModule from "../../../../../../functions/randoms/get-random-value-from-enum";
+
+jest.mock(`../../../../../logger/services/chalk.service`);
+
+describe(`DiscordMessageCommandCookieService`, (): void => {
+  let service: DiscordMessageCommandCookieService;
+  let coreEventService: CoreEventService;
+  let loggerService: LoggerService;
+  let discordSoniaService: DiscordSoniaService;
+  let discordMessageConfigService: DiscordMessageConfigService;
+
+  beforeEach((): void => {
+    coreEventService = CoreEventService.getInstance();
+    loggerService = LoggerService.getInstance();
+    discordSoniaService = DiscordSoniaService.getInstance();
+    discordMessageConfigService = DiscordMessageConfigService.getInstance();
+  });
+
+  describe(`getInstance()`, (): void => {
+    it(`should create a DiscordMessageCommandCookie service`, (): void => {
+      expect.assertions(1);
+
+      service = DiscordMessageCommandCookieService.getInstance();
+
+      expect(service).toStrictEqual(
+        expect.any(DiscordMessageCommandCookieService)
+      );
+    });
+
+    it(`should return the created DiscordMessageCommandCookie service`, (): void => {
+      expect.assertions(1);
+
+      const result = DiscordMessageCommandCookieService.getInstance();
+
+      expect(result).toStrictEqual(service);
+    });
+  });
+
+  describe(`constructor()`, (): void => {
+    let coreEventServiceNotifyServiceCreatedSpy: jest.SpyInstance;
+
+    beforeEach((): void => {
+      coreEventServiceNotifyServiceCreatedSpy = jest
+        .spyOn(coreEventService, `notifyServiceCreated`)
+        .mockImplementation();
+    });
+
+    it(`should notify the DiscordMessageCommandCookie service creation`, (): void => {
+      expect.assertions(2);
+
+      service = new DiscordMessageCommandCookieService();
+
+      expect(coreEventServiceNotifyServiceCreatedSpy).toHaveBeenCalledTimes(1);
+      expect(coreEventServiceNotifyServiceCreatedSpy).toHaveBeenCalledWith(
+        ServiceNameEnum.DISCORD_MESSAGE_COMMAND_COOKIE_SERVICE
+      );
+    });
+  });
+
+  describe(`handleResponse()`, (): void => {
+    let anyDiscordMessage: AnyDiscordMessage;
+
+    let loggerServiceDebugSpy: jest.SpyInstance;
+    let discordSoniaServiceGetCorporationMessageEmbedAuthorSpy: jest.SpyInstance;
+    let discordMessageConfigServiceGetMessageCommandCookieImageColorSpy: jest.SpyInstance;
+    let discordSoniaServiceGetImageUrlSpy: jest.SpyInstance;
+    let discordMessageConfigServiceGetMessageCommandCookieImageUrlSpy: jest.SpyInstance;
+    let getRandomValueFromEnumSpy: jest.SpyInstance;
+
+    beforeEach((): void => {
+      anyDiscordMessage = createMock<AnyDiscordMessage>({
+        id: `dummy-id`,
+      });
+
+      loggerServiceDebugSpy = jest.spyOn(loggerService, `debug`);
+      discordSoniaServiceGetCorporationMessageEmbedAuthorSpy = jest.spyOn(
+        discordSoniaService,
+        `getCorporationMessageEmbedAuthor`
+      );
+      discordMessageConfigServiceGetMessageCommandCookieImageColorSpy = jest.spyOn(
+        discordMessageConfigService,
+        `getMessageCommandCookieImageColor`
+      );
+      discordSoniaServiceGetImageUrlSpy = jest.spyOn(
+        discordSoniaService,
+        `getImageUrl`
+      );
+      discordMessageConfigServiceGetMessageCommandCookieImageUrlSpy = jest.spyOn(
+        discordMessageConfigService,
+        `getMessageCommandCookieImageUrl`
+      );
+      getRandomValueFromEnumSpy = jest.spyOn(
+        GetRandomValueFromEnumModule,
+        `getRandomValueFromEnum`
+      );
+    });
+
+    it(`should log about the command`, (): void => {
+      expect.assertions(2);
+
+      service.handleResponse(anyDiscordMessage);
+
+      expect(loggerServiceDebugSpy).toHaveBeenCalledTimes(1);
+      expect(loggerServiceDebugSpy).toHaveBeenCalledWith({
+        context: `DiscordMessageCommandCookieService`,
+        extendedContext: true,
+        message: `context-[dummy-id] text-cookie command detected`,
+      } as ILoggerLog);
+    });
+
+    it(`should return a Discord message response embed with an author`, (): void => {
+      expect.assertions(1);
+      const messageEmbedAuthor: MessageEmbedAuthor = createMock<
+        MessageEmbedAuthor
+      >();
+      discordSoniaServiceGetCorporationMessageEmbedAuthorSpy.mockReturnValue(
+        messageEmbedAuthor
+      );
+
+      const result: unknown = service.handleResponse(anyDiscordMessage);
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(result.options.embed.author).toStrictEqual(messageEmbedAuthor);
+    });
+
+    it(`should return a Discord message response embed with a color`, (): void => {
+      expect.assertions(1);
+      discordMessageConfigServiceGetMessageCommandCookieImageColorSpy.mockReturnValue(
+        ColorEnum.CANDY
+      );
+
+      const result: unknown = service.handleResponse(anyDiscordMessage);
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(result.options.embed.color).toStrictEqual(ColorEnum.CANDY);
+    });
+
+    it(`should return a Discord message response embed with a description`, (): void => {
+      expect.assertions(1);
+      getRandomValueFromEnumSpy.mockReturnValue(
+        DiscordMessageCommandCookieDescriptionEnum.CHUCK_NORRIS_CAN_NOT_BEAT_ME
+      );
+
+      const result: unknown = service.handleResponse(anyDiscordMessage);
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(result.options.embed.description).toStrictEqual(
+        `Chuck Norris can't beat me.`
+      );
+    });
+
+    describe(`when the description can not be found inside the random cookie descriptions`, (): void => {
+      beforeEach((): void => {
+        getRandomValueFromEnumSpy.mockReturnValue(undefined);
+      });
+
+      it(`should return a Discord message response embed with a description`, (): void => {
+        expect.assertions(1);
+
+        const result: unknown = service.handleResponse(anyDiscordMessage);
+
+        // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+        // @ts-ignore
+        expect(result.options.embed.description).toStrictEqual(`Yes.`);
+      });
+    });
+
+    it(`should return a Discord message response embed with a footer containing an icon and a text`, (): void => {
+      expect.assertions(1);
+      discordSoniaServiceGetImageUrlSpy.mockReturnValue(`dummy-image-url`);
+
+      const result: unknown = service.handleResponse(anyDiscordMessage);
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(result.options.embed.footer).toStrictEqual({
+        iconURL: `dummy-image-url`,
+        text: `Bon appétit`,
+      } as MessageEmbedFooter);
+    });
+
+    describe(`when the Sonia image url is null`, (): void => {
+      beforeEach((): void => {
+        discordSoniaServiceGetImageUrlSpy.mockReturnValue(null);
+      });
+
+      it(`should return a Discord message response embed with a footer but without an icon`, (): void => {
+        expect.assertions(1);
+
+        const result: unknown = service.handleResponse(anyDiscordMessage);
+
+        // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+        // @ts-ignore
+        expect(result.options.embed.footer).toStrictEqual({
+          iconURL: undefined,
+          text: `Bon appétit`,
+        } as MessageEmbedFooter);
+      });
+    });
+
+    describe(`when the Sonia image url is "image-url"`, (): void => {
+      beforeEach((): void => {
+        discordSoniaServiceGetImageUrlSpy.mockReturnValue(`image-url`);
+      });
+
+      it(`should return a Discord message response embed with a footer containing an icon and a text`, (): void => {
+        expect.assertions(1);
+
+        const result: unknown = service.handleResponse(anyDiscordMessage);
+
+        // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+        // @ts-ignore
+        expect(result.options.embed.footer).toStrictEqual({
+          iconURL: `image-url`,
+          text: `Bon appétit`,
+        } as MessageEmbedFooter);
+      });
+    });
+
+    it(`should return a Discord message response embed with a thumbnail`, (): void => {
+      expect.assertions(1);
+      discordMessageConfigServiceGetMessageCommandCookieImageUrlSpy.mockReturnValue(
+        IconEnum.ARTIFICIAL_INTELLIGENCE
+      );
+
+      const result: unknown = service.handleResponse(anyDiscordMessage);
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(result.options.embed.thumbnail).toStrictEqual({
+        url: IconEnum.ARTIFICIAL_INTELLIGENCE,
+      } as MessageEmbedThumbnail);
+    });
+
+    it(`should return a Discord message response embed with a timestamp`, (): void => {
+      expect.assertions(2);
+
+      const result: unknown = service.handleResponse(anyDiscordMessage);
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(moment(result.options.embed.timestamp).isValid()).toStrictEqual(
+        true
+      );
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(moment(result.options.embed.timestamp).fromNow()).toStrictEqual(
+        `a few seconds ago`
+      );
+    });
+
+    it(`should return a Discord message response embed with a title`, (): void => {
+      expect.assertions(1);
+      getRandomValueFromEnumSpy.mockReturnValue(
+        DiscordMessageCommandCookieTitleEnum.COOKIE_DELIVERY
+      );
+
+      const result: unknown = service.handleResponse(anyDiscordMessage);
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(result.options.embed.title).toStrictEqual(`Cookie delivery!`);
+    });
+
+    describe(`when the title can not be found inside the random cookie titles`, (): void => {
+      beforeEach((): void => {
+        getRandomValueFromEnumSpy.mockReturnValue(undefined);
+      });
+
+      it(`should return a Discord message response embed with a title`, (): void => {
+        expect.assertions(1);
+
+        const result: unknown = service.handleResponse(anyDiscordMessage);
+
+        // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+        // @ts-ignore
+        expect(result.options.embed.title).toStrictEqual(`Cookies!`);
+      });
+    });
+
+    it(`should return a Discord message response splitted`, (): void => {
+      expect.assertions(1);
+
+      const result: unknown = service.handleResponse(anyDiscordMessage);
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(result.options.split).toStrictEqual(true);
+    });
+
+    it(`should return a Discord message response without a response text`, (): void => {
+      expect.assertions(1);
+
+      const result = service.handleResponse(anyDiscordMessage);
+
+      expect(result.response).toStrictEqual(``);
+    });
+  });
+});
