+++ src/features/discord/messages/services/command/discord-message-command.service.ts
@@ -5,10 +5,10 @@ import { DiscordMessageCommandEnum } from "../../enums/command/discord-message-c
 import { IDiscordMessageResponse } from "../../interfaces/discord-message-response";
 import { AnyDiscordMessage } from "../../types/any-discord-message";
 import { DiscordMessageConfigService } from "../config/discord-message-config.service";
-import { DiscordMessageCommandCookieService } from "./discord-message-command-cookie.service";
-import { DiscordMessageCommandErrorService } from "./discord-message-command-error.service";
-import { DiscordMessageCommandHelpService } from "./discord-message-command-help.service";
-import { DiscordMessageCommandVersionService } from "./discord-message-command-version.service";
+import { DiscordMessageCommandCookieService } from "./cookie/discord-message-command-cookie.service";
+import { DiscordMessageCommandErrorService } from "./error/discord-message-command-error.service";
+import { DiscordMessageCommandHelpService } from "./help/discord-message-command-help.service";
+import { DiscordMessageCommandVersionService } from "./version/discord-message-command-version.service";
 import { DiscordMessageContentService } from "../discord-message-content.service";
 
 export class DiscordMessageCommandService extends AbstractService {
