+++ src/features/discord/messages/services/command/error/discord-message-command-error.service.ts
@@ -7,13 +7,13 @@ import {
 } from "discord.js";
 import _ from "lodash";
 import moment from "moment-timezone";
-import { AbstractService } from "../../../../../classes/abstract.service";
-import { ServiceNameEnum } from "../../../../../enums/service-name.enum";
-import { LoggerService } from "../../../../logger/services/logger.service";
-import { DiscordSoniaService } from "../../../users/services/discord-sonia.service";
-import { IDiscordMessageResponse } from "../../interfaces/discord-message-response";
-import { AnyDiscordMessage } from "../../types/any-discord-message";
-import { DiscordMessageConfigService } from "../config/discord-message-config.service";
+import { AbstractService } from "../../../../../../classes/abstract.service";
+import { ServiceNameEnum } from "../../../../../../enums/service-name.enum";
+import { LoggerService } from "../../../../../logger/services/logger.service";
+import { DiscordSoniaService } from "../../../../users/services/discord-sonia.service";
+import { IDiscordMessageResponse } from "../../../interfaces/discord-message-response";
+import { AnyDiscordMessage } from "../../../types/any-discord-message";
+import { DiscordMessageConfigService } from "../../config/discord-message-config.service";
 
 export class DiscordMessageCommandErrorService extends AbstractService {
   private static _instance: DiscordMessageCommandErrorService;
