+++ src/features/discord/messages/services/command/version/discord-message-command-version.service.spec.ts
@@ -6,48 +6,56 @@ import {
 } from "discord.js";
 import moment from "moment-timezone";
 import { createMock } from "ts-auto-mock";
-import { ColorEnum } from "../../../../../enums/color.enum";
-import { IconEnum } from "../../../../../enums/icon.enum";
-import { ServiceNameEnum } from "../../../../../enums/service-name.enum";
-import { CoreEventService } from "../../../../core/services/core-event.service";
-import { ILoggerLog } from "../../../../logger/interfaces/logger-log";
-import { LoggerService } from "../../../../logger/services/logger.service";
-import { DiscordSoniaService } from "../../../users/services/discord-sonia.service";
-import { AnyDiscordMessage } from "../../types/any-discord-message";
-import { DiscordMessageConfigService } from "../config/discord-message-config.service";
-import { DiscordMessageCommandHelpService } from "./discord-message-command-help.service";
+import { ColorEnum } from "../../../../../../enums/color.enum";
+import { IconEnum } from "../../../../../../enums/icon.enum";
+import { ServiceNameEnum } from "../../../../../../enums/service-name.enum";
+import { AppProductionStateEnum } from "../../../../../app/enums/app-production-state.enum";
+import { AppConfigQueryService } from "../../../../../app/services/config/app-config-query.service";
+import { AppConfigService } from "../../../../../app/services/config/app-config.service";
+import { CoreEventService } from "../../../../../core/services/core-event.service";
+import { ILoggerLog } from "../../../../../logger/interfaces/logger-log";
+import { LoggerService } from "../../../../../logger/services/logger.service";
+import { DiscordSoniaEmotionalStateEnum } from "../../../../users/enums/discord-sonia-emotional-state.enum";
+import { DiscordSoniaService } from "../../../../users/services/discord-sonia.service";
+import { AnyDiscordMessage } from "../../../types/any-discord-message";
+import { DiscordMessageConfigService } from "../../config/discord-message-config.service";
+import { DiscordMessageCommandVersionService } from "./discord-message-command-version.service";
 
-jest.mock(`../../../../logger/services/chalk.service`);
+jest.mock(`../../../../../logger/services/chalk.service`);
 
-describe(`DiscordMessageCommandHelpService`, (): void => {
-  let service: DiscordMessageCommandHelpService;
+describe(`DiscordMessageCommandVersionService`, (): void => {
+  let service: DiscordMessageCommandVersionService;
   let coreEventService: CoreEventService;
   let loggerService: LoggerService;
   let discordSoniaService: DiscordSoniaService;
   let discordMessageConfigService: DiscordMessageConfigService;
+  let appConfigQueryService: AppConfigQueryService;
+  let appConfigService: AppConfigService;
 
   beforeEach((): void => {
     coreEventService = CoreEventService.getInstance();
     loggerService = LoggerService.getInstance();
     discordSoniaService = DiscordSoniaService.getInstance();
     discordMessageConfigService = DiscordMessageConfigService.getInstance();
+    appConfigQueryService = AppConfigQueryService.getInstance();
+    appConfigService = AppConfigService.getInstance();
   });
 
   describe(`getInstance()`, (): void => {
-    it(`should create a DiscordMessageCommandHelp service`, (): void => {
+    it(`should create a DiscordMessageCommandVersion service`, (): void => {
       expect.assertions(1);
 
-      service = DiscordMessageCommandHelpService.getInstance();
+      service = DiscordMessageCommandVersionService.getInstance();
 
       expect(service).toStrictEqual(
-        expect.any(DiscordMessageCommandHelpService)
+        expect.any(DiscordMessageCommandVersionService)
       );
     });
 
-    it(`should return the created DiscordMessageCommandHelp service`, (): void => {
+    it(`should return the created DiscordMessageCommandVersion service`, (): void => {
       expect.assertions(1);
 
-      const result = DiscordMessageCommandHelpService.getInstance();
+      const result = DiscordMessageCommandVersionService.getInstance();
 
       expect(result).toStrictEqual(service);
     });
@@ -62,14 +70,14 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
         .mockImplementation();
     });
 
-    it(`should notify the DiscordMessageCommandHelp service creation`, (): void => {
+    it(`should notify the DiscordMessageCommandVersion service creation`, (): void => {
       expect.assertions(2);
 
-      service = new DiscordMessageCommandHelpService();
+      service = new DiscordMessageCommandVersionService();
 
       expect(coreEventServiceNotifyServiceCreatedSpy).toHaveBeenCalledTimes(1);
       expect(coreEventServiceNotifyServiceCreatedSpy).toHaveBeenCalledWith(
-        ServiceNameEnum.DISCORD_MESSAGE_COMMAND_HELP_SERVICE
+        ServiceNameEnum.DISCORD_MESSAGE_COMMAND_VERSION_SERVICE
       );
     });
   });
@@ -79,9 +87,17 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
 
     let loggerServiceDebugSpy: jest.SpyInstance;
     let discordSoniaServiceGetCorporationMessageEmbedAuthorSpy: jest.SpyInstance;
-    let discordMessageConfigServiceGetMessageCommandHelpImageColorSpy: jest.SpyInstance;
+    let discordMessageConfigServiceGetMessageCommandVersionImageColorSpy: jest.SpyInstance;
     let discordSoniaServiceGetImageUrlSpy: jest.SpyInstance;
-    let discordMessageConfigServiceGetMessageCommandHelpImageUrlSpy: jest.SpyInstance;
+    let appConfigQueryServiceGetTotalReleaseCountHumanizedSpy: jest.SpyInstance;
+    let discordMessageConfigServiceGetMessageCommandVersionImageUrlSpy: jest.SpyInstance;
+    let discordSoniaServiceGetFullNameSpy: jest.SpyInstance;
+    let appConfigServiceGetVersionSpy: jest.SpyInstance;
+    let appConfigQueryServiceGetReleaseDateHumanizedSpy: jest.SpyInstance;
+    let appConfigQueryServiceGetInitializationDateHumanizedSpy: jest.SpyInstance;
+    let appConfigServiceGetReleaseNotesSpy: jest.SpyInstance;
+    let appConfigQueryServiceGetProductionStateHumanizedSpy: jest.SpyInstance;
+    let discordSoniaServiceGetEmotionalStateSpy: jest.SpyInstance;
 
     beforeEach((): void => {
       anyDiscordMessage = createMock<AnyDiscordMessage>({
@@ -93,17 +109,49 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
         discordSoniaService,
         `getCorporationMessageEmbedAuthor`
       );
-      discordMessageConfigServiceGetMessageCommandHelpImageColorSpy = jest.spyOn(
+      discordMessageConfigServiceGetMessageCommandVersionImageColorSpy = jest.spyOn(
         discordMessageConfigService,
-        `getMessageCommandHelpImageColor`
+        `getMessageCommandVersionImageColor`
       );
       discordSoniaServiceGetImageUrlSpy = jest.spyOn(
         discordSoniaService,
         `getImageUrl`
       );
-      discordMessageConfigServiceGetMessageCommandHelpImageUrlSpy = jest.spyOn(
+      appConfigQueryServiceGetTotalReleaseCountHumanizedSpy = jest.spyOn(
+        appConfigQueryService,
+        `getTotalReleaseCountHumanized`
+      );
+      discordMessageConfigServiceGetMessageCommandVersionImageUrlSpy = jest.spyOn(
         discordMessageConfigService,
-        `getMessageCommandHelpImageUrl`
+        `getMessageCommandVersionImageUrl`
+      );
+      discordSoniaServiceGetFullNameSpy = jest.spyOn(
+        discordSoniaService,
+        `getFullName`
+      );
+      appConfigServiceGetVersionSpy = jest.spyOn(
+        appConfigService,
+        `getVersion`
+      );
+      appConfigQueryServiceGetReleaseDateHumanizedSpy = jest.spyOn(
+        appConfigQueryService,
+        `getReleaseDateHumanized`
+      );
+      appConfigQueryServiceGetInitializationDateHumanizedSpy = jest.spyOn(
+        appConfigQueryService,
+        `getInitializationDateHumanized`
+      );
+      appConfigServiceGetReleaseNotesSpy = jest.spyOn(
+        appConfigService,
+        `getReleaseNotes`
+      );
+      appConfigQueryServiceGetProductionStateHumanizedSpy = jest.spyOn(
+        appConfigQueryService,
+        `getProductionStateHumanized`
+      );
+      discordSoniaServiceGetEmotionalStateSpy = jest.spyOn(
+        discordSoniaService,
+        `getEmotionalState`
       );
     });
 
@@ -114,9 +162,9 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
 
       expect(loggerServiceDebugSpy).toHaveBeenCalledTimes(1);
       expect(loggerServiceDebugSpy).toHaveBeenCalledWith({
-        context: `DiscordMessageCommandHelpService`,
+        context: `DiscordMessageCommandVersionService`,
         extendedContext: true,
-        message: `context-[dummy-id] text-help command detected`,
+        message: `context-[dummy-id] text-version command detected`,
       } as ILoggerLog);
     });
 
@@ -138,7 +186,7 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
 
     it(`should return a Discord message response embed with a color`, (): void => {
       expect.assertions(1);
-      discordMessageConfigServiceGetMessageCommandHelpImageColorSpy.mockReturnValue(
+      discordMessageConfigServiceGetMessageCommandVersionImageColorSpy.mockReturnValue(
         ColorEnum.CANDY
       );
 
@@ -149,100 +197,118 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
       expect(result.options.embed.color).toStrictEqual(ColorEnum.CANDY);
     });
 
-    it(`should return a Discord message response embed with a description`, (): void => {
+    it(`should return a Discord message response embed with 6 fields`, (): void => {
       expect.assertions(1);
 
       const result: unknown = service.handleResponse(anyDiscordMessage);
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
       // @ts-ignore
-      expect(result.options.embed.description)
-        .toStrictEqual(`Below is the complete list of commands.
-    You can either use *--*, *!* or *$* as prefix to run a command.`);
+      expect(result.options.embed.fields).toHaveLength(6);
     });
 
-    it(`should return a Discord message response embed with 5 fields`, (): void => {
+    it(`should return a Discord message response embed with an application version field`, (): void => {
       expect.assertions(1);
-
-      const result: unknown = service.handleResponse(anyDiscordMessage);
-
-      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
-      // @ts-ignore
-      expect(result.options.embed.fields).toHaveLength(5);
-    });
-
-    it(`should return a Discord message response embed with a version field`, (): void => {
-      expect.assertions(1);
+      appConfigServiceGetVersionSpy.mockReturnValue(`8`);
 
       const result: unknown = service.handleResponse(anyDiscordMessage);
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
       // @ts-ignore
       expect(result.options.embed.fields[0]).toStrictEqual({
-        name: `Version (*!version* or *!v*)`,
-        value: `Display my current application version.`,
+        name: `My age`,
+        value: `[8](https://github.com/Sonia-corporation/il-est-midi-discord/releases/tag/8)`,
       } as EmbedFieldData);
     });
 
-    it(`should return a Discord message response embed with an error field`, (): void => {
+    it(`should return a Discord message response embed with a release date field`, (): void => {
       expect.assertions(1);
+      appConfigQueryServiceGetReleaseDateHumanizedSpy.mockReturnValue(
+        `dummy-release-date-humanized`
+      );
 
       const result: unknown = service.handleResponse(anyDiscordMessage);
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
       // @ts-ignore
       expect(result.options.embed.fields[1]).toStrictEqual({
-        name: `Error (*!error* or *!bug*)`,
-        value: `Create a bug in my core system.
-      Do not do this one, of course!`,
+        inline: true,
+        name: `My last birthday`,
+        value: `dummy-release-date-humanized`,
       } as EmbedFieldData);
     });
 
-    it(`should return a Discord message response embed with a help field`, (): void => {
+    it(`should return a Discord message response embed with a initialization date field`, (): void => {
       expect.assertions(1);
+      appConfigQueryServiceGetInitializationDateHumanizedSpy.mockReturnValue(
+        `dummy-initialization-date-humanized`
+      );
 
       const result: unknown = service.handleResponse(anyDiscordMessage);
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
       // @ts-ignore
       expect(result.options.embed.fields[2]).toStrictEqual({
-        name: `Help (*!help* or *!h*)`,
-        value: `Ask for my help, it is obvious!
-      And maybe I will, who knows?`,
+        inline: true,
+        name: `The last time I woken up`,
+        value: `dummy-initialization-date-humanized`,
       } as EmbedFieldData);
     });
 
-    it(`should return a Discord message response embed with a cookie field`, (): void => {
+    it(`should return a Discord message response embed with a release notes field`, (): void => {
       expect.assertions(1);
+      appConfigServiceGetReleaseNotesSpy.mockReturnValue(`dummy-release-notes`);
 
       const result: unknown = service.handleResponse(anyDiscordMessage);
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
       // @ts-ignore
       expect(result.options.embed.fields[3]).toStrictEqual({
-        name: `Cookie (*!cookie* or *!c*)`,
-        value: `Because I am good, life gave me cookies.
-      Now it is my turn to give you some.`,
+        name: `My birthday card`,
+        value: `dummy-release-notes\n\nCheckout all my [birthday cards](https://github.com/Sonia-corporation/il-est-midi-discord/blob/master/CHANGELOG.md).`,
       } as EmbedFieldData);
     });
 
-    it(`should return a Discord message response embed with a more help field`, (): void => {
+    it(`should return a Discord message response embed with a status field`, (): void => {
       expect.assertions(1);
+      appConfigQueryServiceGetProductionStateHumanizedSpy.mockReturnValue(
+        AppProductionStateEnum.DEVELOPMENT
+      );
 
       const result: unknown = service.handleResponse(anyDiscordMessage);
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
       // @ts-ignore
       expect(result.options.embed.fields[4]).toStrictEqual({
-        name: `Further help`,
-        value: `You can also checkout the [readme](https://github.com/Sonia-corporation/il-est-midi-discord/blob/master/README.md).
-      It contains more information about how I work.`,
+        inline: true,
+        name: `My location`,
+        value: `Running in development`,
+      } as EmbedFieldData);
+    });
+
+    it(`should return a Discord message response embed with an emotional state field`, (): void => {
+      expect.assertions(1);
+      discordSoniaServiceGetEmotionalStateSpy.mockReturnValue(
+        DiscordSoniaEmotionalStateEnum.AGITATED
+      );
+
+      const result: unknown = service.handleResponse(anyDiscordMessage);
+
+      // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+      // @ts-ignore
+      expect(result.options.embed.fields[5]).toStrictEqual({
+        inline: true,
+        name: `My emotional state`,
+        value: `Agitated`,
       } as EmbedFieldData);
     });
 
     it(`should return a Discord message response embed with a footer containing an icon and a text`, (): void => {
       expect.assertions(1);
       discordSoniaServiceGetImageUrlSpy.mockReturnValue(`dummy-image-url`);
+      appConfigQueryServiceGetTotalReleaseCountHumanizedSpy.mockReturnValue(
+        `8 versions`
+      );
 
       const result: unknown = service.handleResponse(anyDiscordMessage);
 
@@ -250,13 +316,16 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
       // @ts-ignore
       expect(result.options.embed.footer).toStrictEqual({
         iconURL: `dummy-image-url`,
-        text: `At your service`,
+        text: `8 versions`,
       } as MessageEmbedFooter);
     });
 
     describe(`when the Sonia image url is null`, (): void => {
       beforeEach((): void => {
         discordSoniaServiceGetImageUrlSpy.mockReturnValue(null);
+        appConfigQueryServiceGetTotalReleaseCountHumanizedSpy.mockReturnValue(
+          `8 versions`
+        );
       });
 
       it(`should return a Discord message response embed with a footer but without an icon`, (): void => {
@@ -268,7 +337,7 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
         // @ts-ignore
         expect(result.options.embed.footer).toStrictEqual({
           iconURL: undefined,
-          text: `At your service`,
+          text: `8 versions`,
         } as MessageEmbedFooter);
       });
     });
@@ -276,6 +345,9 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
     describe(`when the Sonia image url is "image-url"`, (): void => {
       beforeEach((): void => {
         discordSoniaServiceGetImageUrlSpy.mockReturnValue(`image-url`);
+        appConfigQueryServiceGetTotalReleaseCountHumanizedSpy.mockReturnValue(
+          `8 versions`
+        );
       });
 
       it(`should return a Discord message response embed with a footer containing an icon and a text`, (): void => {
@@ -287,14 +359,14 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
         // @ts-ignore
         expect(result.options.embed.footer).toStrictEqual({
           iconURL: `image-url`,
-          text: `At your service`,
+          text: `8 versions`,
         } as MessageEmbedFooter);
       });
     });
 
     it(`should return a Discord message response embed with a thumbnail`, (): void => {
       expect.assertions(1);
-      discordMessageConfigServiceGetMessageCommandHelpImageUrlSpy.mockReturnValue(
+      discordMessageConfigServiceGetMessageCommandVersionImageUrlSpy.mockReturnValue(
         IconEnum.ARTIFICIAL_INTELLIGENCE
       );
 
@@ -326,13 +398,14 @@ describe(`DiscordMessageCommandHelpService`, (): void => {
 
     it(`should return a Discord message response embed with a title`, (): void => {
       expect.assertions(1);
+      discordSoniaServiceGetFullNameSpy.mockReturnValue(`dummy-full-name`);
 
       const result: unknown = service.handleResponse(anyDiscordMessage);
 
       // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
       // @ts-ignore
       expect(result.options.embed.title).toStrictEqual(
-        `So, you need my help? Cool.`
+        `dummy-full-name version`
       );
     });
