+++ src/features/discord/messages/services/command/version/discord-message-command-version.service.ts
@@ -7,18 +7,18 @@ import {
 } from "discord.js";
 import _ from "lodash";
 import moment from "moment-timezone";
-import { AbstractService } from "../../../../../classes/abstract.service";
-import { ServiceNameEnum } from "../../../../../enums/service-name.enum";
-import { ellipsis } from "../../../../../functions/formatters/ellipsis";
-import { AppProductionStateEnum } from "../../../../app/enums/app-production-state.enum";
-import { AppConfigQueryService } from "../../../../app/services/config/app-config-query.service";
-import { AppConfigService } from "../../../../app/services/config/app-config.service";
-import { LoggerService } from "../../../../logger/services/logger.service";
-import { DiscordSoniaEmotionalStateEnum } from "../../../users/enums/discord-sonia-emotional-state.enum";
-import { DiscordSoniaService } from "../../../users/services/discord-sonia.service";
-import { IDiscordMessageResponse } from "../../interfaces/discord-message-response";
-import { AnyDiscordMessage } from "../../types/any-discord-message";
-import { DiscordMessageConfigService } from "../config/discord-message-config.service";
+import { AbstractService } from "../../../../../../classes/abstract.service";
+import { ServiceNameEnum } from "../../../../../../enums/service-name.enum";
+import { ellipsis } from "../../../../../../functions/formatters/ellipsis";
+import { AppProductionStateEnum } from "../../../../../app/enums/app-production-state.enum";
+import { AppConfigQueryService } from "../../../../../app/services/config/app-config-query.service";
+import { AppConfigService } from "../../../../../app/services/config/app-config.service";
+import { LoggerService } from "../../../../../logger/services/logger.service";
+import { DiscordSoniaEmotionalStateEnum } from "../../../../users/enums/discord-sonia-emotional-state.enum";
+import { DiscordSoniaService } from "../../../../users/services/discord-sonia.service";
+import { IDiscordMessageResponse } from "../../../interfaces/discord-message-response";
+import { AnyDiscordMessage } from "../../../types/any-discord-message";
+import { DiscordMessageConfigService } from "../../config/discord-message-config.service";
 
 export class DiscordMessageCommandVersionService extends AbstractService {
   private static _instance: DiscordMessageCommandVersionService;
